+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Allow super to be called within state-driven behaviors
+
 ## 1.2.0 / 2013-03-30
 
 * Allow multiple whitelisted / blacklisted :to states when definining transitions